# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KuruOps API
  version: "0.1.0"
  description: >
    REST API for KuruOps, a single-tenant SOC alert/incident-management
    platform (Go backend, PostgreSQL, React frontend). This document covers
    every route mounted by internal/httpserver/router.go: the unauthenticated
    /auth/** login flows, /healthz and /metrics, and the JWT-protected
    /api/v1/** surface (alerts, incidents, playbooks, tags, uploads,
    dashboard, live event stream, and admin-only /settings/** configuration).
    See the repository root README for product context, deployment, and the
    default-admin-first-login flow; see TROUBLESHOOTING.md for operational
    issues. This spec favors completeness of shape (paths, params, request/
    response bodies, status codes) over exhaustive field-level validation
    constraints.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/dashboard/stats:
    get:
      tags: [Dashboard]
      summary: KPI cards, trend charts, and breakdowns
      description: Not gated by resourceAccess -- blends alert and incident KPIs into one summary.
      parameters:
        - name: alertSeverity
          in: query
          schema: { $ref: '#/components/schemas/Severity' }
        - name: alertStatus
          in: query
          schema: { $ref: '#/components/schemas/AlertStatus' }
        - name: alertSource
          in: query
          schema: { type: string }
        - name: alertTag
          in: query
          schema: { type: string }
        - name: assignedAnalystId
          in: query
          schema: { type: string, format: uuid }
        - name: incidentSeverity
          in: query
          schema: { $ref: '#/components/schemas/Severity' }
        - name: incidentTag
          in: query
          schema: { type: string }
        - name: commanderId
          in: query
          schema: { type: string, format: uuid }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: until
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardStats' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/dashboard/activity:
    get:
      tags: [Dashboard]
      summary: Recent Activity feed
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
        - name: kind
          in: query
          description: Filter to "alert" or "incident" activity; omit for both.
          schema: { type: string, enum: [alert, incident] }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: until
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Activity events, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ActivityEvent' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
...
  /api/v1/settings/escalation-policies:
    get:
      tags: [Settings - Escalation Policies]
      summary: List escalation policies
      responses:
        "200":
          description: Escalation policies
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EscalationPolicy' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Settings - Escalation Policies]
      summary: Create or update the escalation policy for a severity
      description: >
        The whole ordered chain of steps is resubmitted and wholesale-replaced
        on every save, same idiom playbook steps use.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                severity: { $ref: '#/components/schemas/Severity' }
                steps:
                  type: array
                  items: { $ref: '#/components/schemas/SaveEscalationStepRequest' }
              required: [severity, steps]
      responses:
        "200":
          description: Policy saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EscalationPolicy' }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v1/settings/escalation-policies/{id}:
    delete:
      tags: [Settings - Escalation Policies]
      summary: Delete an escalation policy
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/settings/escalation-policies/test:
    post:
      tags: [Settings - Escalation Policies]
      summary: Send a real test notification through one step of an already-saved chain
      description: >
        Lets an admin confirm a destination (and, for webhook, a custom
        payload template with the analyst placeholders) actually works
        without waiting for a real alert to escalate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                severity: { $ref: '#/components/schemas/Severity' }
                stepPosition:
                  type: integer
                  description: 0-indexed, matching EscalationStep.position.
              required: [severity, stepPosition]
      responses:
        "200":
          description: Test notification sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'

  # ------------------------------------------------- Settings: Audit Export
  # ------------------------------------------------------ Settings: Audit Log
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Required on every /api/v1/** route. /auth/** is intentionally
        unauthenticated (that's how a session is obtained in the first
        place), and /healthz and /metrics are unauthenticated for
        infrastructure health checks and scraping. A caller whose
        mustChangePassword flag is still set can reach nothing under
        /api/v1/** except POST /api/v1/account/change-password (see
        RequirePasswordChanged), receiving 403 on everything else until the
        password is changed.
...
  responses:
    BadRequest:
      description: Malformed request body, path, or query parameter
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Unauthorized:
      description: Missing, malformed, or expired bearer token
      content:
        text/plain:
          schema:
            type: string
            description: >-
              Plain-text error message, issued by the auth middleware via
              http.Error -- not the handler-level JSON error shape used
              elsewhere in this API.
    Forbidden:
      description: Authenticated, but the caller's role/resourceAccess doesn't permit this action (or the account still must change its password)
      content:
        text/plain:
          schema:
            type: string
            description: >-
              Plain-text error message, issued by the auth middleware via
              http.Error -- not the handler-level JSON error shape used
              elsewhere in this API.
...
    InternalError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: Standard handler-level error shape, written by handlers.writeError.
      properties:
        error:
          type: string
      required: [error]
...
    Severity:
      type: string
      enum: [critical, high, medium, low, informational]
    AlertStatus:
      type: string
      enum: [open, investigating, escalated, closed]
...
    ActorType:
      type: string
      enum: [user, system, ai]
...
    EscalationChannelType:
      type: string
      enum: [pagerduty, slack, webhook]
...
    EscalationStep:
      type: object
      description: >
        One link in an escalation policy's chain. position is 0-indexed and
        dense (0..len(steps)-1). destinationSecretRef is never serialized --
        resolved server-side only.
      properties:
        id: { type: string, format: uuid }
        policyId: { type: string, format: uuid }
        position: { type: integer }
        scheduleId:
          type: string
          format: uuid
          description: Which on-call schedule this step resolves the on-call analyst against -- not necessarily the tenant's default schedule.
        scheduleName: { type: string }
        delayMinutes:
          type: integer
          description: How long to wait after the previous event before this step fires.
        channelType: { $ref: '#/components/schemas/EscalationChannelType' }
        webhookPayloadTemplate:
          type: string
          nullable: true
          description: Only meaningful when channelType is webhook; omitted/null means send the default fixed JSON payload shape.
    SaveEscalationStepRequest:
      type: object
      properties:
        scheduleId: { type: string, format: uuid }
        delayMinutes: { type: integer }
        channelType: { $ref: '#/components/schemas/EscalationChannelType' }
        destination:
          type: string
          description: Plaintext destination (e.g. Slack webhook URL, PagerDuty routing key); empty on update keeps the existing one for that position.
        webhookPayloadTemplate:
          type: string
          description: Only applies when channelType is webhook; empty means send the default fixed payload shape.
      required: [scheduleId, delayMinutes, channelType]
    EscalationPolicy:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        severity: { $ref: '#/components/schemas/Severity' }
        steps:
          type: array
          items: { $ref: '#/components/schemas/EscalationStep' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AlertTrendPoint:
      type: object
      properties:
        day: { type: string }
        alertCount: { type: integer }
        avgMttrSeconds: { type: number, nullable: true }
    IncidentTrendPoint:
      type: object
      properties:
        day: { type: string }
        incidentCount: { type: integer }
    NamedCount:
      type: object
      description: A breakdown row keyed by identity; a nil id is the "unassigned"/"no commander yet" bucket.
      properties:
        id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        count: { type: integer }
    DashboardStats:
      type: object
      properties:
        openAlerts: { type: integer }
        criticalAlerts: { type: integer }
        highAlerts: { type: integer }
        activeIncidents: { type: integer }
        slaBreachedCount: { type: integer }
        p1OpenCount: { type: integer }
        incidentAvgMttaSeconds: { type: number, nullable: true }
        incidentAvgMttrSeconds: { type: number, nullable: true }
        alertAvgMttaSeconds: { type: number, nullable: true }
        alertAvgMttrSeconds: { type: number, nullable: true }
        alertTrend:
          type: array
          items: { $ref: '#/components/schemas/AlertTrendPoint' }
        incidentTrend:
          type: array
          items: { $ref: '#/components/schemas/IncidentTrendPoint' }
        alertsBySeverity:
          type: object
          additionalProperties: { type: integer }
        alertStatusDistribution:
          type: object
          additionalProperties: { type: integer }
        incidentsByPriority:
          type: object
          additionalProperties: { type: integer }
        incidentsByPhase:
          type: object
          additionalProperties: { type: integer }
        alertsByAnalyst:
          type: array
          items: { $ref: '#/components/schemas/NamedCount' }
        incidentsByCommander:
          type: array
          items: { $ref: '#/components/schemas/NamedCount' }
    ActivityEvent:
      type: object
      description: One row of the Dashboard's Recent Activity feed -- a union of alert_events and incident_events, newest first.
      properties:
        kind:
          type: string
          enum: [alert, incident]
        contextId: { type: string, format: uuid }
        contextTitle: { type: string }
        eventType: { type: string }
        actorType: { $ref: '#/components/schemas/ActorType' }
        actorId: { type: string, format: uuid, nullable: true }
        data:
          description: Event-type-specific JSON payload.
        createdAt: { type: string, format: date-time }
...